+++ pkgs/intl4x/test/ecma/list_format_test.dart
@@ -15,7 +15,7 @@ import '../utils.dart';
 void main() {
   group('List style options', () {
     final list = ['A', 'B', 'C'];
-    final intl = Intl(defaultLocale: 'en_US');
+    final intl = Intl(locale: const Locale(language: 'en', region: 'US'));
     testWithFormatting('long', () {
       final listFormat =
           intl.listFormat(const ListFormatOptions(style: ListStyle.long));
@@ -35,7 +35,7 @@ void main() {
 
   group('List type options', () {
     final list = ['A', 'B', 'C'];
-    final intl = Intl(defaultLocale: 'en_US');
+    final intl = Intl(locale: const Locale(language: 'en', region: 'US'));
     testWithFormatting('long', () {
       final listFormat =
           intl.listFormat(const ListFormatOptions(type: Type.conjunction));
@@ -55,7 +55,9 @@ void main() {
 
   group('List style and type combinations', () {
     final list = ['A', 'B', 'C'];
-    final intl = Intl(ecmaPolicy: const AlwaysEcma(), defaultLocale: 'en_US');
+    final intl = Intl(
+        ecmaPolicy: const AlwaysEcma(),
+        locale: const Locale(language: 'en', region: 'US'));
     testWithFormatting('long', () {
       final formatter = intl.listFormat(const ListFormatOptions(
           style: ListStyle.narrow, type: Type.conjunction));