+++ pkgs/intl4x/test/icu4x/list_format_test.dart
@@ -14,14 +14,14 @@ import '../utils.dart';
 void main() {
   final list = ['A', 'B', 'C'];
   test('Does not compare in tests', () {
-    final locale = 'de_DE';
-    final listFormatGerman = Intl(defaultLocale: locale)
+    final locale = const Locale(language: 'de', region: 'DE');
+    final listFormatGerman = Intl(locale: locale)
         .listFormat(const ListFormatOptions(style: ListStyle.long));
     expect(listFormatGerman.format(list), '${list.join(', ')}//$locale');
   });
 
   testWithFormatting('long', () {
-    final intl = Intl(defaultLocale: 'en_US');
+    final intl = Intl(locale: const Locale(language: 'en', region: 'US'));
     final listFormat =
         intl.listFormat(const ListFormatOptions(style: ListStyle.long));
     expect(() => listFormat.format(list), throwsA(isA<UnimplementedError>()));